Move const where they are used
