+++ crates/interpreter/src/interpreter/stack.rs
@@ -1,9 +1,12 @@
 use crate::{
-    primitives::{B256, STACK_LIMIT, U256},
+    primitives::{B256, U256},
     InstructionResult,
 };
 use alloc::vec::Vec;
 use core::fmt;
+
+/// EVM interpreter stack limit.
+pub const STACK_LIMIT: usize = 1024;
 
 /// EVM stack.
 #[derive(Clone, Debug, Eq, PartialEq)]
