+++ crates/revm/src/lib.rs
@@ -28,7 +28,7 @@ pub use db::{
 
 pub use db::{Database, DatabaseCommit, InMemoryDB};
 pub use evm::{evm_inner, new, EVM};
-pub use evm_impl::{EVMData, EVMImpl, Transact};
+pub use evm_impl::{EVMData, EVMImpl, Transact, CALL_STACK_LIMIT};
 pub use journaled_state::{is_precompile, JournalCheckpoint, JournalEntry, JournaledState};
 
 // reexport `revm_precompiles`
